Extract types to remove circular imports
