+++ src/sqlfluff/core/parser/grammar/base.py
@@ -11,11 +11,11 @@ from typing import (
     Tuple,
     Any,
     cast,
-    FrozenSet,
 )
 from uuid import uuid4
 
 from sqlfluff.core.errors import SQLParseError
+from sqlfluff.core.parser.grammar.types import SimpleHintType, MatchableType
 from sqlfluff.core.string_helpers import curtail_string
 
 from sqlfluff.core.parser.segments import BaseSegment, BracketedSegment, allow_ephemeral
@@ -31,8 +31,6 @@ from sqlfluff.core.parser.context import ParseContext
 from sqlfluff.core.parser.parsers import BaseParser
 
 # Either a Matchable (a grammar or parser) or a Segment CLASS
-MatchableType = Union[Matchable, Type[BaseSegment]]
-SimpleHintType = Optional[Tuple[FrozenSet[str], FrozenSet[str]]]
 
 if TYPE_CHECKING:
     from sqlfluff.core.dialects.base import Dialect  # pragma: no cover
